Move more items out of onmt, update attention to output a prediction over vocabulary
